refactor: break out att init code
